extracted join
