+++ ormar/fields/__init__.py
@@ -1,6 +1,6 @@
 from ormar.fields.base import BaseField
 from ormar.fields.foreign_key import ForeignKey
-from ormar.fields.many_to_many import ManyToMany
+from ormar.fields.many_to_many import ManyToMany, ManyToManyField
 from ormar.fields.model_fields import (
     BigInteger,
     Boolean,
@@ -29,5 +29,6 @@ __all__ = [
     "Time",
     "ForeignKey",
     "ManyToMany",
+    "ManyToManyField",
     "BaseField",
 ]
